Overview: ---
Add's support for Batch API Calls ( one API call sent to Eventbrite to process multiple API calls with one returned response ).

See https://www.eventbriteapi.com/v3/batch/ for details.

Weirdness:
---

The `batch` endpoint does not appear to support `application/json` encoding as the rest of the API does.  This appears to be because a batch parameter is required that contains JSON ( stringified ).  So there is a check added if the `resource` is `batch` and replaces the `Content-type` header to `application/x-www-form-urlencoded` which is what their API seemed to use.

Sample Code:
---

```js
this.client = eventbrite({
   token: config.get('eventbrite.personal_token')
});

var params = {
  "batch": [
    {
      "method": "GET",
      "relative_url": "events/search/?venue.city=Portland&venue.region=OR&venue.country=US"
    },
    {
      "method": "GET",
      "relative_url": "events/search/?venue.city=Portland&venue.region=OR&venue.country=US&page=2"
    },
    {
      "method": "GET",
      "relative_url": "events/search/?venue.city=Portland&venue.region=OR&venue.country=US&page=3"
    }
  ]
};

self.client.batch(params, function(error, data) {
  if (error) {
    // do something with error
  } else {
    // data has the combined output of all API calls in an array, one item for each API call
  }
});
```